Audit Customer removed: MySql not allow to use cycle trigger.
It isn't possible to update audit table with a trigger, because this hypothetical trigger (T1) is fired by an event on Customer, and T1 may causes an insert into audit table.
But if an user is in the audit table, it is marked as insolvet in Customer table, so when T1 does an insert another trigger (T2) have to start for update Customer.
We have a cycle T1->T2->T1